# backend/openapi.yaml
openapi: 3.0.3
info:
  title: V PATHing Rewards API
  version: 1.0.0
  description: Cashback rewards platform API
servers:
  - url: https://vpathrewards.store/api
    description: Production
security:
  - bearerAuth: []
  - cookieAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Merchant:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        website_url:
          type: string
          nullable: true
        logo_url:
          type: string
          nullable: true
...
paths:
...
  /merchants:
    get:
      operationId: listMerchants
      summary: List merchants
      security: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
            maxLength: 200
        - name: category
          in: query
          schema:
            type: string
            maxLength: 100
        - name: minCashback
          in: query
          schema:
            type: number
            minimum: 0
            maximum: 100
        - name: sort
          in: query
          schema:
            type: string
            enum: [name, cashback_asc, cashback_desc, newest]
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Merchant list
  /merchants/{id}:
    get:
      operationId: getMerchant
      summary: Get merchant by ID
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Merchant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Merchant'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /merchants/{id}/offers:
    get:
      operationId: getMerchantOffers
      summary: Get offers for a merchant
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Offers
...